Improve usability and output of script

+++ app/services/tasks/solr_migration_service.rb
@@ -1,4 +1,5 @@
 require 'time'
+require 'ruby-progressbar'
 
 module Tasks
   # Service for reindexing objects from one solr instance to another
@@ -51,22 +52,31 @@ module Tasks
     end
 
     # Trigger indexing of all objects listed in the provided file
-    def reindex(id_list_file)
+    def reindex(id_list_file, clean_index)
+      id_total = File.foreach(id_list_file).inject(0) {|c, line| c+1}
+
       progress_logger = Migrate::Services::ProgressTracker.new(progress_log_path(id_list_file))
-      Blacklight.default_index.connection.delete_by_query('*:*')
-      Blacklight.default_index.connection.commit
-      puts "Deleting everything"
-      sleep(10)
+      if clean_index
+        puts "**** Clearing index ****"
+        Blacklight.default_index.connection.delete_by_query('*:*')
+        Blacklight.default_index.connection.commit
+      end
+
+      progressbar = ProgressBar.create(:total => id_total,
+          :length => 80,
+          :format => "%E |%b\u{15E7}%i| %p%% (%c / %C \u{0394}%R)",
+          :progress_mark => ' ',
+          :remainder_mark => "\u{FF65}")
+
       # Read input file
       id_file = File.new(id_list_file)
       id_file.each_line do |id_line|
         id = id_line.chomp
-        puts "Indexing #{id}"
+        # puts "Indexing #{id}"
         object = ActiveFedora::Base.find(id)
         # Must use update_index instead of going to SolrService.add in order to trigger NestingCollection behaviors
         object.update_index
-        # ActiveFedora::SolrService.add(object.to_solr, commit: true)
-        # puts "Got object #{object.to_solr}: #{outcome}"
+        progressbar.increment
         progress_logger.add_entry(id)
       end
     end
